+++ qubic-rpc/target/debug/examples/v2_api_demo-c9c286d750e3ac91.d
@@ -1,0 +1,5 @@
+/Users/apple/qanchor-cli/qubic-rpc/target/debug/examples/v2_api_demo-c9c286d750e3ac91: examples/v2_api_demo.rs
+
+/Users/apple/qanchor-cli/qubic-rpc/target/debug/examples/v2_api_demo-c9c286d750e3ac91.d: examples/v2_api_demo.rs
+
+examples/v2_api_demo.rs: